Go away variable size nazi ....

+++ sensors.h
@@ -8,6 +8,6 @@
 #ifndef SENSORS_H_
 #define SENSORS_H_
 
-void fillSensorData(int array[8]);
+void fillSensorData(short array[8]);
 
 #endif
